Revert "Merge r198995 - bmalloc: segregate small and large objects again, and allocate more objects on the small path"
This reverts commit 51a2f842da169184e7f9592b64ee30676d349e1f.

+++ Source/bmalloc/bmalloc/BumpAllocator.h
@@ -51,8 +51,8 @@ public:
 
 private:
     char* m_ptr;
-    unsigned m_size;
-    unsigned m_remaining;
+    unsigned short m_size;
+    unsigned short m_remaining;
 };
 
 inline BumpAllocator::BumpAllocator()
